Address warnings affecting the consolidated seq library.
- Do away with implicit narrowing conversions, which yield warnings
  under recent macOS and FreeBSD toolchains, mainly through a
  combination of type harmonization (including adjustments to some
  internal functions' signatures) and explicit casting.
-- CDense_seg: Replace the usual sequence of consistency assertions
   with calls to a new ASSERT_CONSISTENCY macro, which does the same
   thing but on a pure size_t basis.
- Use string::find when simply looking for a single case-sensitive
  character.
- CReverseComplementHelper: Make fully empty, which has always been
  legitimate.
- SeqTable_multi_data.cpp (sx_Round): Correctly detect values that are
  just too large for 64-bit integer types.

JIRA: CXX-3707.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@99064 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/objects/seq/seq_align_mapper_base.hpp
@@ -114,7 +114,7 @@ struct NCBI_SEQ_EXPORT SAlignment_Segment
     int       m_GroupIdx;       ///< Group of segments (e.g. an exon)
     /// Group of scores. Set when several segments share the same set of
     /// scores. Currently used only for sparse-segs. -1 = unassigned.
-    int       m_ScoresGroupIdx;
+    ssize_t   m_ScoresGroupIdx;
 
     // Used only for spliced exon parts to indicate their type.
     TPartType m_PartType;
@@ -272,8 +272,8 @@ private:
     // is converted to to disc.
     void x_ConvToDstDisc(CRef<CSeq_align>& dst) const;
     // Get the next part of the disc align - see x_ConvToDstDisc.
-    int x_GetPartialDenseg(CRef<CSeq_align>& dst,
-                           int               start_seg) const;
+    ssize_t x_GetPartialDenseg(CRef<CSeq_align>& dst,
+                               size_t            start_seg) const;
 
     // Collect exons from a single sub-alignment.
     void x_GetDstSplicedSubAlign(CSpliced_seg&                 spliced,
